isolated git logic from server script; added versionless service
