+++ .gitignore
@@ -1,1 +1,2 @@
 server.log
+/mozilla-pipeline-schemas/**
